Restructure and prepare for Dockerization.
