--- .env.example
@@ -1,3 +1,0 @@
-BOT_TOKEN=
-LOG_DIR=./logs/log.txt
-DEFAULT_PREFIX=/
