+++ .gitignore
@@ -1,3 +1,5 @@
 log.txt
 .env
-node_modules/
+azuri.env
+/src/persist/guilds.json
+/src/node_modules
