+++ azuri.sample.env
@@ -1,0 +1,3 @@
+# Enter your Discord bot token.
+# Generate one here: https://discord.com/developers/applications
+BOT_TOKEN=
