+++ build_entrypoint.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+npm ci
+
+exec "$@"
