+++ src/commands/join.js
