+++ src/commands/nowPlaying.js
