+++ src/commands/stats.js
